Move tests to test_util and fix line lengths
